added viewholder, working on sunrise/sunset
