+++ app/build.gradle.kts
@@ -52,6 +52,7 @@ android {
 }
 
 dependencies {
+    implementation("com.github.bumptech.glide:glide:4.14.2")
     implementation("androidx.navigation:navigation-fragment:2.2.2")
     implementation("androidx.navigation:navigation-ui:2.2.2")
     implementation("androidx.cardview:cardview:1.0.0")
