+++ app/src/main/java/com/capstone/weatherapp/WeatherApiResponse.kt
@@ -12,21 +12,6 @@ data class City(
     val id: Int,
     val name: String
 )
-data class SingleCityResponse(
-    val coord: Coord,
-    val weather: List<Weather>,
-    val base: String,
-    val main: Main,
-    val visibility: Int,
-    val wind: Wind,
-    val clouds: Clouds,
-    val dt: Int,
-    val sys: Sys,
-    val timezone: Int,
-    val id: Int,
-    val name: String,
-    val cod: Int
-)
 data class Coord(
     val lon: Double,
     val lat: Double
